Better split and structure the annotation processing stuff
