+++ aot-maven-plugin/pom.xml
@@ -29,7 +29,6 @@
     <dependency>
       <groupId>com.github.javaparser</groupId>
       <artifactId>javaparser-core</artifactId>
-      <version>${javaparser.version}</version>
     </dependency>
     <dependency>
       <groupId>org.apache.maven</groupId>
