--- function-annotations/pom.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>com.dylibso.chicory</groupId>
-    <artifactId>chicory</artifactId>
-    <version>999-SNAPSHOT</version>
-    <relativePath>../pom.xml</relativePath>
-  </parent>
-  <artifactId>function-annotations</artifactId>
-  <packaging>jar</packaging>
-  <name>Chicory - Function annotations</name>
-  <description>Experimental annotations for host functions</description>
-
-</project>
