--- function-annotations/src/main/java/com/dylibso/chicory/function/annotations/Buffer.java
@@ -1,10 +1,0 @@
-package com.dylibso.chicory.function.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.PARAMETER)
-public @interface Buffer {}
