--- function-annotations/src/main/java/com/dylibso/chicory/function/annotations/HostModule.java
@@ -1,12 +1,0 @@
-package com.dylibso.chicory.function.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface HostModule {
-    String value();
-}
