+++ host-module/annotations/src/main/java/com/dylibso/chicory/hostmodule/annotations/WasmExport.java
@@ -1,0 +1,12 @@
+package com.dylibso.chicory.hostmodule.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface WasmExport {
+    String value() default "";
+}
