+++ host-module/it/src/it/with-imports/invoker.properties
@@ -1,0 +1,1 @@
+invoker.goals=test
