+++ host-module/pom.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>com.dylibso.chicory</groupId>
+    <artifactId>chicory</artifactId>
+    <version>999-SNAPSHOT</version>
+    <relativePath>../pom.xml</relativePath>
+  </parent>
+  <artifactId>host-module-parent</artifactId>
+  <packaging>pom</packaging>
+  <name>Chicory - Host Module - Parent</name>
+  <description>Parent module for the Host Module machinery</description>
+
+  <modules>
+    <module>annotations</module>
+    <module>it</module>
+    <module>processor</module>
+  </modules>
+
+</project>
